Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
numpy==1.26.4
tensorflow==2.17.0
opencv-python==4.8.1.78
werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==21.2.0
Pillow==10.0.1
scikit-learn==1.5.0
matplotlib==3.8.0
